refactor: simplify WebRTC configuration
